Change Player to Ecto Schema
Using a GenServer to store the JSON entries has shown not to be a good
idea. The drawback of using this approach is that it would be hard to do
some operations (as paginating and sorting) without overloading the
memory in some time.

We're now relying on Ecto and its functions to validate and store the
entries instead of implementing everything.