Change player's table into a live component
Now, instead of having the rows/columns hard-coded into the main live
template, we've split the table logic into live components.

The table now is more generic, and can be reused for different schemas.

A module was created to hydrate the table component with the necessary
fields to create their rows.

* Created a Player UI helper for populating the table columns
* Table split into a more generic live component